bitcoin-demo-v4/bitcoin-demo-v3/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-anthropic>=0.0.2
faiss-cpu>=1.7.4
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
gitpython>=3.1.40
tqdm>=4.66.1 